// src/types.ts
export enum DownloadMode {
  AIStudio = 'AIStudio',
  ModelFinder = 'ModelFinder',
  NewModels = 'NewModels',
  LoraFinder = 'LoraFinder',
  HuggingFace = 'HuggingFace',
  CustomUrl = 'CustomUrl',
}

export enum DownloadDestination {
  Local = 'Local',
  GoogleDrive = 'GoogleDrive',
}

export enum ModelFamily {
  Gemini25Flash = 'gemini-2.5-flash',
}